feat(api/applications): allows array of regions associations with case (BOAS-129)
