+++ apps/api/src/server/applications/application/__tests__/create-application.test.js
@@ -18,6 +18,12 @@ const findUniqueZoomLevelStub = sinon.stub();
 findUniqueZoomLevelStub.withArgs({ where: { name: 'some-unknown-map-zoom-level' } }).returns(null);
 findUniqueZoomLevelStub.withArgs({ where: { name: 'some-known-map-zoom-level' } }).returns({});
 
+const findUniqueRegionStub = sinon.stub();
+
+findUniqueRegionStub.withArgs({ where: { name: 'region1' } }).returns({});
+findUniqueRegionStub.withArgs({ where: { name: 'region2' } }).returns({});
+findUniqueRegionStub.withArgs({ where: { name: 'some-unknown-region' } }).returns(null);
+
 test.before('set up mocks', () => {
 	sinon.stub(databaseConnector, 'case').get(() => {
 		return { create: createStub };
@@ -29,6 +35,10 @@ test.before('set up mocks', () => {
 
 	sinon.stub(databaseConnector, 'zoomLevel').get(() => {
 		return { findUnique: findUniqueZoomLevelStub };
+	});
+
+	sinon.stub(databaseConnector, 'region').get(() => {
+		return { findUnique: findUniqueRegionStub };
 	});
 });
 
@@ -120,6 +130,7 @@ test('creates new application when all possible details provided', async (t) => 
 		geographicalInformation: {
 			mapZoomLevelName: 'some-known-map-zoom-level',
 			locationDescription: 'location description',
+			regionNames: ['region1', 'region2'],
 			gridReference: {
 				easting: '123456',
 				northing: '987654'
@@ -144,7 +155,13 @@ test('creates new application when all possible details provided', async (t) => 
 					locationDescription: 'location description',
 					firstNotifiedAt: new Date(123),
 					submissionAt: new Date(1_689_262_804_000),
-					subSector: { connect: { name: 'some_sub_sector' } }
+					subSector: { connect: { name: 'some_sub_sector' } },
+					regions: {
+						create: [
+							{ region: { connect: { name: 'region1' } } },
+							{ region: { connect: { name: 'region2' } } }
+						]
+					}
 				}
 			},
 			serviceCustomer: {
@@ -223,7 +240,8 @@ test('returns error if any validated values are invalid', async (t) => {
 				easting: '123',
 				northing: '12345879'
 			},
-			mapZoomLevelName: 'some-unknown-map-zoom-level'
+			mapZoomLevelName: 'some-unknown-map-zoom-level',
+			regionNames: ['some-unknown-region']
 		},
 		applicant: {
 			email: 'not a real email',
@@ -248,6 +266,7 @@ test('returns error if any validated values are invalid', async (t) => {
 			'geographicalInformation.gridReference.easting': 'Easting must be integer with 6 digits',
 			'geographicalInformation.gridReference.northing': 'Northing must be integer with 6 digits',
 			'geographicalInformation.mapZoomLevelName': 'Must be a valid map zoom level',
+			'geographicalInformation.regionNames': 'Unknown region',
 			'keyDates.firstNotifiedDate': 'First notified date must be in the past',
 			'keyDates.submissionDate': 'Submission date must be in the future',
 			subSectorName: 'Must be existing sub-sector'
